added test functions for geospatial resolution detection using synthetic and real data
